feat: recator AssessmentLegend
