+++ app/components/AnalystDashboard/index.tsx
@@ -4,3 +4,4 @@ export { default as AssessmentAssignmentTable } from './AssessmentAssignmentTabl
 export { default as AssessmentLead } from './AssessmentLead';
 export { default as DashboardTabs } from './DashboardTabs';
 export { default as TableTabs } from './TableTabs';
+export { default as AssessmentLegend } from './AssessmentLegend';
